i don't want the mview file anymore
